requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

opencv-python>=4.8.0
pillow>=10.1.0
timm>=0.9.0

shap>=0.43.0
lime>=0.2.0.1

elasticsearch>=8.11.0
langdetect>=1.0.9
langchain>=0.1.0
fasttext-wheel>=0.9.2

requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0